refactor(ckd): restructure appid in its own submodule
